changed stuff and made laser not hit the same thing twice

+++ SifiGame/SifiGame/bin/Debug/AppX/js/projectile_collision.js
@@ -44,46 +44,46 @@
             for (playercounter = 0; playercounter < playercollection.count() ; playercounter++) {
                 if (collisiondetection1.testcollisionenemy(projectilecollection.array[projectilecounter], playercollection.array[playercounter]) == 1) {
                     if (playercollection.array[playercounter].type == "enemy") {
-                        if (projectilecollection.array[projectilecounter].canexplode)
-                        {
+                        if (projectilecollection.array[projectilecounter].checkifinarray(playercounter) == 0) {
+                            projectilecollection.array[projectilecounter].addhitarray(playercounter);
 
- 
-                            projectilecollection.array[projectilecounter].explode();
-                            projectilecollection.delete(projectilecounter);
-                        }
-                        else
-                        {
-                            playercollection.array[playercounter].health = playercollection.array[playercounter].health - projectilecollection.array[projectilecounter].damagemultiplier;
-                            if (playercollection.array[projectilecollection.array[projectilecounter].pindex].type == "player") {
-                                game2.addmoney(playercollection.array[projectilecollection.array[projectilecounter].pindex].index, 10);
+                            if (projectilecollection.array[projectilecounter].canexplode) {
+
+
+                                projectilecollection.array[projectilecounter].explode();
+                                projectilecollection.delete(projectilecounter);
                             }
-                            else
-                            {
-                                if (playercollection.array[projectilecollection.array[projectilecounter].pindex].type == "drone")
-                                {
-                                    game2.addmoney(playercollection.array[projectilecollection.array[projectilecounter].pindex].parentindex, 10);
+                            else {
+                                playercollection.array[playercounter].health = playercollection.array[playercounter].health - projectilecollection.array[projectilecounter].damagemultiplier;
+                                if (playercollection.array[projectilecollection.array[projectilecounter].pindex].type == "player") {
+                                    game2.addmoney(playercollection.array[projectilecollection.array[projectilecounter].pindex].index, 10);
                                 }
+                                else {
+                                    if (playercollection.array[projectilecollection.array[projectilecounter].pindex].type == "drone") {
+                                        game2.addmoney(playercollection.array[projectilecollection.array[projectilecounter].pindex].parentindex, 10);
+                                    }
+                                }
+                                if ((Math.random() * 100) + 1 > 50) {
+                                    var velmultiplyertestx = -1;
+                                }
+                                else {
+                                    var velmultiplyertestx = 1;
+                                }
+                                if ((Math.random() * 100) + 1 > 50) {
+                                    var velmultiplyertesty = -1;
+                                }
+                                else {
+                                    var velmultiplyertesty = 1;
+                                }
+                                floating_numbercollection.add(playercollection.array[playercounter].x, playercollection.array[playercounter].y, ((Math.random() * 2) + 0.1) * velmultiplyertestx, ((Math.random() * 2) + 0.1) * velmultiplyertesty, projectilecollection.array[projectilecounter].damagemultiplier, 15, "rgb(200,75,75)")
                             }
-                            if ((Math.random() * 100) + 1 > 50) {
-                                var velmultiplyertestx = -1;
-                            }
-                            else {
-                                var velmultiplyertestx = 1;
-                            }
-                            if ((Math.random() * 100) + 1 > 50) {
-                                var velmultiplyertesty = -1;
-                            }
-                            else {
-                                var velmultiplyertesty = 1;
-                            }
-                            floating_numbercollection.add(playercollection.array[playercounter].x, playercollection.array[playercounter].y, ((Math.random() * 2) + 0.1) * velmultiplyertestx, ((Math.random() * 2) + 0.1) * velmultiplyertesty, projectilecollection.array[projectilecounter].damagemultiplier, 15, "rgb(200,75,75)")
-                        }
 
-                        if (projectilecollection.array[projectilecounter].armourpiercing == 0) {
-                            projectilecollection.delete(projectilecounter);
-                        }
-						break;
+                            if (projectilecollection.array[projectilecounter].armourpiercing == 0) {
+                                projectilecollection.delete(projectilecounter);
+                            }
+                            break;
 
+                        }
                     }
 
                 }
